+++ R/create_hub_schema.R
@@ -1,6 +1,6 @@
 #' Create a Hub arrow schema
 #'
-#' Create an arrow schema from a the `tasks.json` config file for using when
+#' Create an arrow schema from a `tasks.json` config file. For use when
 #' opening an arrow dataset.
 #' @param config_tasks a list version of the content's of a hub's `tasks.json`
 #' config file created using function [read_config()].
@@ -8,6 +8,14 @@
 #' @param partitions a named list specifying the arrow data types of any
 #' partitioning column. Only relevant for opening `"parquet"` or `"arrow"` format datasets.
 #' Ignored for all other formats.
+#' @param output_type_id_datatype character string. One of `"auto"`, `"character"`,
+#' `"double"`, `"integer"`, `"logical"`, `"Date"`. Defaults to `"auto"` indicating
+#' that `output_type_id` will be determined automatically from the `tasks.json`
+#' config file. Other data type values can be used to override automatic determination.
+#' Note that attempting to coerce `output_type_id` to a data type that is not possible
+#' (e.g. trying to coerce to `"double"` when the data contains `"character"` values)
+#' will likely result in an error or potentially unexpected behaviour so use with
+#' care.
 #'
 #' @return an arrow schema object that can be used to define column datatypes when
 #' opening model output data. If the schema is being created for a `parquet`/`arrow` hub,
@@ -22,8 +30,13 @@
 #' schema_parquet <- create_hub_schema(config_tasks, format = "parquet")
 create_hub_schema <- function(config_tasks,
                               format = c("csv", "parquet", "arrow"),
-                              partitions = list(model_id = arrow::utf8())) {
+                              partitions = list(model_id = arrow::utf8()),
+                              output_type_id_datatype = c("auto", "character",
+                                                          "double", "integer",
+                                                          "logical", "Date")) {
   format <- rlang::arg_match(format)
+  output_type_id_datatype <- rlang::arg_match(output_type_id_datatype)
+
   if (format == "csv") {
     partitions <- NULL
   }
@@ -49,9 +62,15 @@ create_hub_schema <- function(config_tasks,
   task_id_arrow_types <- arrow_datatypes[task_id_types] %>%
     stats::setNames(task_id_names)
 
+  if (output_type_id_datatype == "auto") {
+    output_type_id_type <- arrow_datatypes[[get_output_type_id_type(config_tasks)]]
+  } else {
+    output_type_id_type <- arrow_datatypes[[output_type_id_datatype]]
+  }
+
   c(task_id_arrow_types,
     output_type = arrow::utf8(),
-    output_type_id = arrow_datatypes[[get_output_type_id_type(config_tasks)]],
+    output_type_id = output_type_id_type,
     value = arrow_datatypes[[get_value_type(config_tasks)]],
     partitions
   ) %>%