+++ tests/testthat/test-hub-connection.R
@@ -87,7 +87,13 @@ test_that("connect_hub file_format override works on local hub", {
   )
 })
 
+test_that("Overriding output_type_id data type works correctly", {
+  con <- connect_hub(hub_path, output_type_id_datatype = "character")
+  expect_equal(con$schema$output_type_id$ToString(),
+               "output_type_id: string")
 
+
+})
 
 
 test_that("connect_model_output works on local model_output_dir", {